+++ engine/mainui/Utils.cpp
@@ -214,7 +214,7 @@ void COM_FileBase ( const char *in, char *out )
 {
 	int len, start, end;
 
-	len = strlen( in );
+	len = (int)strlen( in );
 	
 	// scan backward for '.'
 	end = len - 1;
