+++ engine/mainui/menus/Controls.cpp
@@ -148,7 +148,7 @@ void CMenuControls::UnbindCommand( const char *command )
 	int i, l;
 	const char *b;
 
-	l = strlen( command );
+	l = (int)strlen( command );
 
 	for( i = 0; i < 256; i++ )
 	{
